backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
idna==3.10
psycopg==3.2.3
python-dotenv==1.0.0
redis==5.2.1
requests==2.32.3
six==1.17.0
spotipy==2.23.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
